workers/base/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Michael Chernyakoff", email = "chernyakoff@gmail.com"},
]
dependencies = [
    "tortoise-orm[accel,asyncpg]>=0.25.1",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "aiopath>=0.7.7",
    "autoflake>=2.3.1",
    "cyclopts>=3.19.0",
    "watchfiles>=1.1.0",
    "python-dateutil>=2.9.0.post0",
    "emojis>=0.7.0",
    "hatchet-sdk>=1.18.1",
    "aiohttp-socks>=0.10.1",
    "astor>=0.8.1",
    "phonenumbers>=9.0.13",
    "telethon>=1.41.2",
    "cryptg>=0.5.1",
    "aioboto3>=15.1.0",
    "pyyaml>=6.0.2",
    "pydantic-settings-yaml>=0.2.0",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.taskipy.tasks]
app = "python -m app"
dev = "python -m app.dev"
autoflake = "autoflake ./app"
cli = "python -m app.cli"


[tool.autoflake]
ignore-init-module-imports = true
remove-all-unused-imports = true
recursive = true
in-place = true

[tool.sync.codegen]
worker_module = "app.worker"
output_dir = "../../api/app/hatchet/base"
client_import = "from app.hatchet.client import hatchet"

[tool.sync.modules]
src = "../../api/app"
dest = "app"
modules = ["common"]


[tool.worker]
name = "base"
slots = 10

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]
